+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [41.3.0] 2026-04-02
+### Added
+- New utility function LForms.Util.validateQuestionnaireOnFHIRServer()
+  to validate a questionnaire against a FHIR server.
+
 ## [41.2.0] 2026-04-02
 ### Added
 - Support for targetConstraint extension.